Decode/convert serde plain enum value using subshape codecs #25
